ad ladder class

+++ core/src/main/java/com/mystudio/gamename/Variable.java
@@ -1,8 +1,8 @@
 package com.mystudio.gamename;
 
 public class Variable {
-    public static float SCREEN_WIDTH = 1500;
-    public static float SCREEN_HEIGHT = 1000;
+    public static int SCREEN_WIDTH = 1500;
+    public static int SCREEN_HEIGHT = 1000;
     public static float GRAVITY = 4;
     public static int MAX_OBJ = 50;
 }
